Backend/LockPad/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
pymongo>=4.6
python-dotenv>=1.0
cryptography>=42.0
django-cors-headers>=4.3
gunicorn>=21.2
